Herramientas/requirements.txt
requests
pandas
xlsxwriter
python-dotenv
apify-client
